Migrated to x86 SFML. Added fullscreen mode.

+++ SFML_Raycasting/GameObject.cs
@@ -4,6 +4,6 @@ namespace Raycasting
 {
     interface GameObject
     {
-        void Draw(RenderWindow app, float horPos, float height);
+        void Draw(RenderWindow app, int horPos, float height);
     }
 }
